[libft] modify ft_strnstr, ft_lstadd_back, and type of ft_strlen.

+++ cursus/libft/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: yekim <yekim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/30 16:01:38 by yekim             #+#    #+#             */
-/*   Updated: 2020/10/05 09:20:01 by yekim            ###   ########.fr       */
+/*   Updated: 2020/10/06 07:41:01 by yekim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 char	*ft_strdup(const char *s)
 {
 	char	*ret;
-	int		len;
+	size_t	len;
 
-	len = (int)ft_strlen((char *)s);
+	len = ft_strlen((char *)s);
 	if (!(ret = (char *)malloc(sizeof(char) * (len + 1))))
 		return (NULL);
 	ft_strlcpy(ret, s, len + 1);
