+++ cursus/libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: yekim <yekim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/29 17:43:57 by yekim             #+#    #+#             */
-/*   Updated: 2020/10/04 19:00:47 by yekim            ###   ########.fr       */
+/*   Updated: 2020/10/06 07:42:12 by yekim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	int	len_dest;
-	int	len_src;
-	int	itr;
+	size_t	len_dest;
+	size_t	len_src;
+	size_t	itr;
 
 	len_dest = 0;
 	len_src = ft_strlen((char *)src);
@@ -25,11 +25,11 @@ size_t	ft_strlcat(char *dest, const char *src, size_t size)
 		++dest;
 		++len_dest;
 	}
-	if (size <= (size_t)len_dest)
-		return ((size_t)len_src + size);
-	itr = (int)size - len_dest + 1;
+	if (size <= len_dest)
+		return (len_src + size);
+	itr = size - len_dest + 1;
 	while (*src && --itr)
 		*(dest++) = *(src++);
 	*dest = '\0';
-	return (size_t)(len_src + len_dest);
+	return (len_src + len_dest);
 }
