+++ cursus/libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: yekim <yekim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/29 17:44:00 by yekim             #+#    #+#             */
-/*   Updated: 2020/10/04 18:26:36 by yekim            ###   ########.fr       */
+/*   Updated: 2020/10/06 07:42:30 by yekim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-	int	ret;
-	int	itr;
+	size_t	ret;
+	size_t	itr;
 
 	ret = 0;
-	itr = (int)size;
+	itr = size;
 	while (*src && --itr)
 	{
 		*(dest++) = *(src++);
